backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
cors==1.0.1
openai>=1.0.0
pyarrow>=14.0.0